Changed beams_data to just live in the beams.hpp file.
Those wanting input handling should include beams_input directly
