+++ src/restruct_poc/InterpolateQPState.hpp
@@ -2,7 +2,7 @@
 
 #include <Kokkos_Core.hpp>
 #include "types.hpp"
-#include "beams_data.hpp"
+#include "beams.hpp"
 
 namespace openturbine {
