--- src/restruct_poc/dynamic_beam.cpp
@@ -1,1 +1,0 @@
-#include "solver.hpp"
