+++ tests/unit_tests/restruct_poc/test_rotor.cpp
@@ -6,6 +6,7 @@
 #include <gtest/gtest.h>
 
 #include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/beams_input.hpp"
 #include "src/restruct_poc/solver.hpp"
 #include "src/restruct_poc/Step.hpp"
 #include "src/restruct_poc/InitializeConstraints.hpp"
