+++ tests/unit_tests/restruct_poc/test_solver.cpp
@@ -3,6 +3,7 @@
 #include <gtest/gtest.h>
 
 #include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/beams_input.hpp"
 #include "src/restruct_poc/solver.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 #include "src/restruct_poc/InitializeConstraints.hpp"
